+++ src/apis/user.ts
@@ -1,0 +1,8 @@
+import { client } from './client'
+
+/**
+ * checkNicknameValid
+ */
+export const checkNicknameValid = (nickname: string) => {
+  return client.post(`/user/nickname`, { nickname })
+}
